compute-demo/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-cli>=4.0.0,<5.0.0
